Account

+++ src/main/java/com/projet/proxy/service/AccountService.java
@@ -1,9 +1,12 @@
 package com.projet.proxy.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.projet.proxy.model.Account;
 import com.projet.proxy.repository.AccountDao;
@@ -21,18 +24,32 @@ public class AccountService implements IAccountService {
 
 	@Override
 	public Account getAccountById(Long id) {
-		return accountDao.findById(id).orElse(null);
+	    Optional<Account> account = accountDao.findById(id);
+	    if (account.isPresent()) {
+	        return account.get();
+	    } else {
+	        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Account not found with id: " + id);
+	    }
 	}
+
 
 	@Override
 	public Account createAccount(Account account) {
-		return accountDao.save(account);
+	    if (account.getId() != null && accountDao.existsById(account.getId())) {
+	        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Account already exists with id: " + account.getId());
+	    }
+	    return accountDao.save(account);
 	}
 
+
 	@Override
-	public Account updateAccount(Account account) {
-		return accountDao.save(account);
+	public Account updateAccount(Account updatedAccount) {
+	    if (!accountDao.existsById(updatedAccount.getId())) {
+	        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Account not found with id: " + updatedAccount.getId());
+	    }
+	    return accountDao.save(updatedAccount);
 	}
+
 
 	@Override
 	public boolean isAccountExists(Long id) {
@@ -41,7 +58,11 @@ public class AccountService implements IAccountService {
 
 	@Override
 	public void deleteAccountById(Long id) {
-		accountDao.deleteById(id);
+	    if (accountDao.existsById(id)) {
+	        accountDao.deleteById(id);
+	    } else {
+	        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Account not found with id: " + id);
+	    }
 	}
 
 }
